backend/requirements.txt
Flask==2.3.2
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
flask-jwt-extended==4.4.4
python-dotenv==1.0.0
passlib==1.7.4
email-validator==1.3.1
itsdangerous==2.1.2
